swagger docs / changed id type from Long to Integer

+++ rest/src/main/java/com/ang/rest/domain/entities/TransactionDetailsId.java
@@ -10,6 +10,6 @@ import java.io.Serializable;
 @AllArgsConstructor
 @EqualsAndHashCode
 public class TransactionDetailsId implements Serializable {
-    private Long transaction;
-    private Long product;
+    private Integer transaction;
+    private Integer product;
 }
